Added feature to only save if Maschine has focus
Because I am using a keystroke to save, I've made a change to detect
whether or not Maschine has focus. If it does it will save, otherwise
it waits until Maschine has focus again. Without this while you have
the script and Maschine running the a [Command}+S keystroke will be
sent to any application that has focus.